Improve server setup script

+++ setup.server.sh
@@ -4,7 +4,7 @@ sudo apt-get update
 sudo apt-get upgrade -y
 
 echo "Installing basic tools"
-sudo apt-get install vim gitk unzip python-setuptools build-essential python-dev libfftw3-dev libasound2-dev zlib1g-dev python-tk gcc -y
+sudo apt-get install vim gitk unzip python-setuptools build-essential python-dev libfftw3-dev libasound2-dev zlib1g-dev python-tk libatlas-dev libatlas-base-dev libblas-dev liblapack-dev gfortran -y
 
 echo "Installing clang"
 wget -O - https://apt.llvm.org/llvm-snapshot.gpg.key | sudo apt-key add -
@@ -16,7 +16,7 @@ sudo update-alternatives --install /usr/bin/clang clang /usr/bin/clang-5.0 1000
 sudo update-alternatives --config clang --skip-auto
 sudo update-alternatives --config clang++ --skip-auto
 
-type cmake
+type cmake  &> /dev/null
 if [ $? -eq 1 ]
 then
     echo "Installing cmake"
@@ -26,17 +26,24 @@ then
     rm -f cmake-3.10.2-Linux-x86_64.tar.gz
 fi
 
-echo "* Downloading ndk 16"
-wget --quiet --show-progress https://dl.google.com/android/repository/android-ndk-r16b-linux-x86_64.zip
-echo "* Extracting ndk 16"
-unzip -q android-ndk-r16b-linux-x86_64.zip
-echo "* Moving ndk 16 to /opt"
-sudo mv android-ndk-r16b /opt
-rm -f android-ndk-r16b-linux-x86_64.zip
+if [ ! -d /opt/android-ndk-r16b ]
+then
+    echo "* Downloading ndk 16"
+    wget --quiet --show-progress https://dl.google.com/android/repository/android-ndk-r16b-linux-x86_64.zip
+    echo "* Extracting ndk 16"
+    unzip -q android-ndk-r16b-linux-x86_64.zip
+    echo "* Moving ndk 16 to /opt"
+    sudo mv android-ndk-r16b /opt
+    rm -f android-ndk-r16b-linux-x86_64.zip
+fi
 
-echo "Adding cmake to PATH..."
-echo "export PATH=/opt/cmake/bin:\$PATH" >> ~/.profile
-source ~/.profile
+grep "export PATH=\/opt\/cmake\/bin\:\$PATH" ~/.profile  &> /dev/null
+if [ $? -eq 1 ]
+then
+    echo "Adding cmake to PATH..."
+    echo "export PATH=/opt/cmake/bin:\$PATH" >> ~/.profile
+    source ~/.profile
+fi
 
 echo "Python installations"
 sudo easy_install pip
@@ -60,34 +67,49 @@ echo "setup git"
 git config --global push.default simple
 git config --global credential.helper cache
 git config --global credential.helper 'cache --timeout=3600'
-echo "checkout Bugatone-Space"
+
 cd
-git clone https://github.com/Bugatone/Bugatone-Space.git
-sudo chown -R ubuntu:ubuntu Bugatone-Space
-sudo -H pip install -e Bugatone-Space
+if [ ! -d Bugatone-Space ]
+then
+    echo "checkout Bugatone-Space"
+    git clone https://github.com/Bugatone/Bugatone-Space.git
+    sudo chown -R ubuntu:ubuntu Bugatone-Space
+    sudo -H pip install -e Bugatone-Space
+fi
 
-echo "checkout mobileproduct"
-git clone https://github.com/Bugatone/mobileproduct.git
-sudo -H pip install -e mobileproduct
+if [ ! -d mobileproduct ]
+then
+    echo "checkout mobileproduct"
+    git clone https://github.com/Bugatone/mobileproduct.git
+    sudo -H pip install -e mobileproduct
+fi
 
-echo "installing git-lfs"
-wget -q --show-progress https://github.com/git-lfs/git-lfs/releases/download/v2.4.2/git-lfs-linux-amd64-2.4.2.tar.gz
-tar -xzf git-lfs-linux-amd64-2.4.2.tar.gz
-sudo -H git-lfs-2.4.2/install.sh
-rm -r git-lfs-2.4.2
-rm git-lfs-linux-amd64-2.4.2.tar.gz
-git lfs install
-echo "checkout test-files"
-git clone https://github.com/Bugatone/test-files.git
+git lfs &> /dev/null
+if [ $? -eq 1 ]
+then
+    echo "installing git-lfs"
+    wget -q --show-progress https://github.com/git-lfs/git-lfs/releases/download/v2.4.2/git-lfs-linux-amd64-2.4.2.tar.gz
+    tar -xzf git-lfs-linux-amd64-2.4.2.tar.gz
+    sudo -H git-lfs-2.4.2/install.sh
+    rm -r git-lfs-2.4.2
+    rm git-lfs-linux-amd64-2.4.2.tar.gz
+    git lfs install
+    echo "checkout test-files"
+    git clone https://github.com/Bugatone/test-files.git
+fi
 
-echo "Updating buga .profile file"
-echo "export BUGATONE_ROOT=/home/ubuntu/Bugatone-Space" >> ~/.profile
-echo "export TEST_FILES_PATH=/home/ubuntu/test-files" >> ~/.profile
-echo "export LD_LIBRARY_PATH=.:/home/ubuntu/Bugatone-Space/lib/linux_x86:\$LD_LIBRARY_PATH" >> ~/.profile
-echo "export PATH=.:/home/ubuntu/Bugatone-Space/bin/linux_x86:\$PATH" >> ~/.profile
-echo "export PYTHONPATH=/home/ubuntu/Bugatone-Space/python" >> ~/.profile
-echo "export PATH=/opt/cmake/bin:\$PATH" >> ~/.profile
-source ~/.profile
+grep "export BUGATONE_ROOT" ~/.profile  &> /dev/null
+if [ $? -eq 1 ]
+then
+    echo "Updating buga .profile file"
+    echo "export BUGATONE_ROOT=/home/ubuntu/Bugatone-Space" >> ~/.profile
+    echo "export TEST_FILES_PATH=/home/ubuntu/test-files" >> ~/.profile
+    echo "export LD_LIBRARY_PATH=.:/home/ubuntu/Bugatone-Space/lib/linux_x86:\$LD_LIBRARY_PATH" >> ~/.profile
+    echo "export PATH=.:/home/ubuntu/Bugatone-Space/bin/linux_x86:\$PATH" >> ~/.profile
+    echo "export PYTHONPATH=/home/ubuntu/Bugatone-Space/python" >> ~/.profile
+    echo "export PATH=/opt/cmake/bin:\$PATH" >> ~/.profile
+    source ~/.profile
+fi
 
 echo "Create bugatone folders..."
 mkdir ~/tmp_dir
@@ -100,4 +122,4 @@ sudo chown -R ubuntu:ubuntu /home/ubuntu
 echo "make Bugatone-Space"
 cd Bugatone-Space
 ./make.sh linux
-.
+
